added if statement to return a message to show if the product exist o…

+++ src/main/java/com/BilBay/bilbay/controllers/ProductController.java
@@ -1,5 +1,6 @@
 package com.BilBay.bilbay.controllers;
 import com.BilBay.bilbay.models.Product;
+import com.BilBay.bilbay.repositories.ProductRepository;
 import com.BilBay.bilbay.services.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -9,6 +10,8 @@ import java.util.List;
 public class ProductController {
     @Autowired
     ProductService productService;
+    @Autowired
+    ProductRepository productRepository;
     public ProductController(ProductService productService) {
         this.productService = productService;
     }
@@ -31,9 +34,13 @@ public class ProductController {
     }
     @DeleteMapping("delete-product/{id}")
     public String deleteById(@PathVariable Long id){
-        productService.deleteProduct(id);
-        return "Product has been deleted.";
+        if (!productRepository.findById(id).isEmpty()){
+            productService.deleteProduct(id);
+            return "Product has been deleted.";
+        }
+        return "Product not found";
     }
+    
     @PutMapping("{id}/{isAvailable}")
     public String updateProductStatus (@PathVariable Long id, @PathVariable boolean isAvailable) {
         productService.updateProductStatus(id, isAvailable);
